Add tests for the library methods. All except 2 pass right now.
The two that don't pass, fail because I have changed what their output
should be.  In the docs, it is shown that the output of
`generate_minor_scale`, given 'C4', is:

    ['C4', 'D4', 'E4', 'F4', 'G4', 'A4', 'A#4']

This is incorrect.  The actual minor scale is:

    ['C4', 'D4', 'Eb4', 'F4', 'G4', 'Ab4', 'Bb4']

The same kind of inconsistency is found in the `generate_minor_triad`
output.  This is not a proper minor triad:

    ['C5', 'D#5', 'G5']

because C -> D# is not a minor third interval, it is an augmented second
interval.  I know, for all practical purposes it will generate the same
tone, but my musical OCD can't stand to see it this way lol!